feat: refactor DetailsView for improved structure and readability; add separate components for header, metrics, attributes, and raw data display.
