+++ packages/core/src/components/DetailsViewAttributesTab.tsx
@@ -1,0 +1,38 @@
+import type { SpanCardType } from "../types/span.ts";
+
+import { CollapsibleSection } from "./CollapsibleSection.tsx";
+import { Input } from "./Input.tsx";
+
+interface AttributesTabProps {
+  data: SpanCardType;
+}
+
+export const DetailsViewAttributesTab = ({ data }: AttributesTabProps) => (
+  <div className="space-y-6">
+    {data.attributes.map((attribute, index) => {
+      const value =
+        attribute.value.stringValue ||
+        attribute.value.intValue ||
+        attribute.value.boolValue?.toString() ||
+        "N/A";
+
+      return (
+        <CollapsibleSection
+          key={`${attribute.key}-${index}`}
+          title={attribute.key}
+          defaultOpen
+        >
+          <Input
+            id={`${data.id}-attribute-${index}`}
+            value={value}
+            disabled
+            readOnly
+            clearable={false}
+            inputClassName="font-mono text-xs"
+            className="w-full"
+          />
+        </CollapsibleSection>
+      );
+    })}
+  </div>
+);
